# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Koji API",
    "description": "Governed Linux control plane API",
    "version": "v0.1.3"
  },
  "servers": [
    {
      "url": "/",
      "description": "Koji daemon"
    }
  ],
...
  "paths": {
...
    "/api/bootstrap": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Bootstrap first user",
        "description": "Creates the first local user only while no users exist.",
        "x-koji-auth-required": false,
        "x-koji-capability": "public-until-first-user",
        "x-koji-csrf-required": false,
        "x-koji-errors": [
          "validation_error",
          "job_conflict",
          "unexpected_response"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Credentials"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthSession"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "description": "Authenticates a user and creates a session plus CSRF token.",
        "x-koji-auth-required": false,
        "x-koji-capability": "public",
        "x-koji-csrf-required": false,
        "x-koji-errors": [
          "unauthenticated",
          "validation_error",
          "unexpected_response"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Credentials"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthSession"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout",
        "description": "Revokes the current session.",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "x-koji-auth-required": true,
        "x-koji-capability": "authenticated",
        "x-koji-csrf-required": true,
        "x-koji-errors": [
          "unauthenticated",
          "csrf_missing_or_invalid",
          "session_expired",
          "unexpected_response"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Status"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/login/magic-token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Magic token login",
        "description": "Consumes a one-time magic token and creates a session.",
        "operationId": "post_api_login_magic_token",
        "x-koji-auth-required": false,
        "x-koji-capability": "public",
        "x-koji-errors": [
          "magic_token_invalid",
          "magic_token_expired",
          "validation_error",
          "unexpected_response"
        ],
        "responses": {
          "200": {
            "description": "AuthSessionResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          },
          "default": {
            "description": "Safe API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicTokenLoginRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "koji_session",
        "description": "HttpOnly session cookie issued by login or bootstrap."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token required for state-changing authenticated requests."
      }
    },
...
    "requestBodies": {
      "Credentials": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CredentialRequest"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "AuthSession": {
        "description": "Authenticated session response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthSessionResponse"
            }
          }
        }
      },
...
      "Status": {
        "description": "Status response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StatusResponse"
            }
          }
        }
      },
...
      "Error": {
        "description": "Safe error response",
        "headers": {
          "X-Request-ID": {
            "description": "Request correlation ID",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "CredentialRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "AuthSessionResponse": {
        "type": "object",
        "required": [
          "authenticated",
          "username",
          "csrfToken"
        ],
        "properties": {
          "authenticated": {
            "type": "boolean"
          },
          "username": {
            "type": "string"
          },
          "csrfToken": {
            "type": "string"
          }
        }
      },
...
      "StatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Safe server error text returned by current API responses."
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "description": "Plain-language client normalized message when available."
          },
          "request_id": {
            "type": "string",
            "description": "Request correlation ID when included in a response body. Current API also returns X-Request-ID."
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "agent_not_implemented",
          "agent_unavailable",
          "csrf_missing_or_invalid",
          "forbidden",
          "identity_user_not_found",
          "job_conflict",
          "job_not_found",
          "magic_token_expired",
          "magic_token_invalid",
          "mutation_disabled",
          "network_error",
          "self_lockout_prevented",
          "service_not_allowlisted",
          "session_expired",
          "unauthenticated",
          "unexpected_response",
          "validation_error"
        ]
      },
      "MagicTokenLoginRequest": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
    }
  }
}